Closureクラス分離
